Instance signer and reorg to signatures
