+++ lootbox/settings.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from bugout.app import Bugout
 
-from .signatures import Signer, BrownieAccountSigner
+from .signatures import Signer, BrownieAccountSigner, InstanceSigner
 
 # Bugout
 BUGOUT_BROOD_URL = os.environ.get("BUGOUT_BROOD_URL", "https://auth.bugout.dev")
@@ -53,9 +53,9 @@ SIGNER_KEYSTORE = os.environ.get("SIGNER_KEYSTORE")
 SIGNER_PASSWORD = os.environ.get("SIGNER_PASSWORD")
 if SIGNER_KEYSTORE is not None and SIGNER_PASSWORD is not None:
     DROP_SIGNER = BrownieAccountSigner(SIGNER_KEYSTORE, SIGNER_PASSWORD)
-
+MOONSTREAM_SIGNING_SERVER_URI = os.environ.get("MOONSTREAM_SIGNING_SERVER_URI", None)
 if DROP_SIGNER is None:
-    raise ValueError("Could not load signer")
+    DROP_SIGNER = InstanceSigner(MOONSTREAM_SIGNING_SERVER_URI)
 
 DROP_DEADLINE_RAW = os.environ.get("DROP_DEADLINE")
 if DROP_DEADLINE_RAW is None:
@@ -82,3 +82,5 @@ if MOONSTREAM_AWS_SIGNER_LAUNCH_TEMPLATE_ID is None:
 MOONSTREAM_AWS_SIGNER_IMAGE_ID = os.environ.get("MOONSTREAM_AWS_SIGNER_IMAGE_ID")
 if MOONSTREAM_AWS_SIGNER_IMAGE_ID is None:
     raise ValueError("MOONSTREAM_AWS_SIGNER_IMAGE_ID environment variable must be set")
+
+MOONSTREAM_AWS_SIGNER_INSTANCE_PORT = 7181
