+++ lootbox/signatures.py
@@ -2,17 +2,73 @@
 Signing and signature verification functionality and interfaces.
 """
 import abc
+import logging
+from typing import Any, List, Optional
 
+import boto3
 from brownie import accounts
 from brownie.network import web3
 from eth_account._utils.signing import sign_message_hash
 import eth_keys
+import requests
 from hexbytes import HexBytes
+
+from . import signing
+from .settings import (
+    AWS_DEFAULT_REGION,
+    MOONSTREAM_AWS_SIGNER_LAUNCH_TEMPLATE_ID,
+    MOONSTREAM_AWS_SIGNER_IMAGE_ID,
+    MOONSTREAM_AWS_SIGNER_INSTANCE_PORT,
+)
+
+logger = logging.getLogger(__name__)
+
+aws_client = boto3.client("ec2", region_name=AWS_DEFAULT_REGION)
+
+
+class AWSDescribeInstancesFail(Exception):
+    """
+    Raised when AWS describe instances command failed.
+    """
+
+
+class AWSRunInstancesFail(Exception):
+    """
+    Raised when AWS run instances command failed.
+    """
+
+
+class AWSTerminateInstancesFail(Exception):
+    """
+    Raised when AWS terminate instances command failed.
+    """
+
+
+class SigningInstancesNotFound(Exception):
+    """
+    Raised when signing instances with the given ids is not found in at AWS.
+    """
+
+
+class SigningInstancesTerminationLimitExceeded(Exception):
+    """
+    Raised when provided several instances to termination.
+    """
+
+
+class SignWithInstanceFail(Exception):
+    """
+    Raised when failed signing of message with instance server.
+    """
 
 
 class Signer:
     @abc.abstractmethod
     def sign_message(self, message):
+        pass
+
+    @abc.abstractmethod
+    def refresh_signer(self):
         pass
 
 
@@ -35,3 +91,147 @@ class BrownieAccountSigner(Signer):
 
 def get_signing_account(raw_message: str, signature: str) -> str:
     return web3.eth.account.recover_message(raw_message, signature=signature)
+
+
+class InstanceSigner(Signer):
+    """
+    AWS instance server signer.
+    """
+
+    def __init__(self, url: Optional[str] = None) -> None:
+        self.current_signer_uri = None
+        if url is not None:
+            self.current_signer_uri = url
+
+    def refresh_signer(self) -> None:
+        try:
+            instances = list_signing_instances()
+        except AWSDescribeInstancesFail:
+            raise AWSDescribeInstancesFail("AWS describe instances command failed")
+        except Exception as err:
+            logger.error(f"AWS describe instances command failed: {err}")
+            raise SignWithInstanceFail("AWS describe instances command failed")
+
+        if len(instances) != 1:
+            raise SignWithInstanceFail("Unsupported number of signing instances")
+
+        self.current_signer_uri = f"http://{instances[0]['private_ip_address']}:{MOONSTREAM_AWS_SIGNER_INSTANCE_PORT}/sign"
+
+    def sign_message(self, message: str):
+        if self.current_signer_uri is None:
+            raise SignWithInstanceFail("Signer url not specified")
+        
+        signed_message = ""
+        try:
+            resp = requests.post(
+                self.current_signer_uri,
+                headers={"Content-Type: application/json"},
+                json={"unsigned_data": message},
+            )
+            resp.raise_for_status()
+            body = resp.json()
+            signed_message = body["signed_data"]
+        except Exception as err:
+            logger.error(f"Failed signing of message with instance server, {err}")
+            raise SignWithInstanceFail("Failed signing of message with instance server")
+
+        return signed_message
+
+
+def list_signing_instances(
+    signing_instances: List[str],
+) -> List[Any]:
+    """
+    Return a list of signing instances with IPs.
+    """
+    described_instances = []
+    try:
+        described_instances_response = aws_client.describe_instances(
+            Filters=[
+                {"Name": "image-id", "Values": [MOONSTREAM_AWS_SIGNER_IMAGE_ID]},
+                {"Name": "tag:Application", "Values": ["signer"]},
+            ],
+            InstanceIds=signing_instances,
+        )
+        for r in described_instances_response["Reservations"]:
+            for i in r["Instances"]:
+                described_instances.append(
+                    {
+                        "instance_id": i["InstanceId"],
+                        "private_ip_address": i["PrivateIpAddress"],
+                    }
+                )
+    except Exception as err:
+        logger.error(f"AWS describe instances command failed: {err}")
+        raise AWSDescribeInstancesFail("AWS describe instances command failed.")
+
+    return described_instances
+
+
+def wakeup_signing_instances(run_confirmed=False, dry_run=True) -> List[str]:
+    """
+    Run new signing instances.
+    """
+    run_instances = []
+    if run_confirmed:
+        try:
+            run_instances_response = aws_client.run_instances(
+                LaunchTemplate={
+                    "LaunchTemplateId": MOONSTREAM_AWS_SIGNER_LAUNCH_TEMPLATE_ID
+                },
+                MinCount=1,
+                MaxCount=1,
+                DryRun=dry_run,
+            )
+            for i in run_instances_response["Instances"]:
+                run_instances.append(i["InstanceId"])
+        except Exception as err:
+            logger.error(f"AWS run instances command failed: {err}")
+            raise AWSRunInstancesFail("AWS run instances command failed")
+
+    return run_instances
+
+
+def sleep_signing_instances(
+    signing_instances: List[str], termination_confirmed=False, dry_run=True
+) -> List[str]:
+    """
+    Fetch, describe, verify signing instances and terminate them.
+    """
+    if len(signing_instances) == 0:
+        raise SigningInstancesNotFound("There are no signing instances to describe")
+
+    described_instances = []
+    try:
+        described_instances_response = list_signing_instances(signing_instances)
+        for i in described_instances_response:
+            described_instances.append(i["instance_id"])
+    except Exception as err:
+        logger.error(f"AWS describe instances command failed: {err}")
+        raise AWSDescribeInstancesFail("AWS describe instances command failed.")
+
+    if len(described_instances) == 0:
+        raise SigningInstancesNotFound(
+            "Signing instances with the given ids is not found in at AWS."
+        )
+    if len(described_instances) > 1:
+        raise SigningInstancesTerminationLimitExceeded(
+            f"Provided {len(described_instances)} instances to termination"
+        )
+
+    terminated_instances = []
+    if termination_confirmed:
+        try:
+            terminated_instances_response = aws_client.terminate_instances(
+                InstanceIds=described_instances,
+                DryRun=dry_run,
+            )
+            for i in terminated_instances_response["TerminatingInstances"]:
+                terminated_instances.append(i["InstanceId"])
+        except Exception as err:
+            logger.error(
+                f"Unable to terminate instance {described_instances}, error: {err}"
+            )
+            raise AWSTerminateInstancesFail("AWS terminate instances command failed")
+
+    return terminated_instances
